use separate controller-runtime cache & address comments
