+++ internal/server/everest.go
@@ -42,7 +42,6 @@ import (
 	"k8s.io/apimachinery/pkg/fields"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/percona/everest/api"
 	"github.com/percona/everest/cmd/config"
@@ -66,7 +65,6 @@ type EverestServer struct {
 	sessionMgr    *session.Manager
 	attemptsStore *RateLimiterMemoryStore
 	handler       handlers.Handler
-	blocklist     session.Blocklist
 	oidcProvider  *oidc.ProviderConfig
 }
 
@@ -95,14 +93,7 @@ func getOIDCProviderConfig(ctx context.Context, kubeClient kubernetes.Kubernetes
 
 // NewEverestServer creates and configures everest API.
 func NewEverestServer(ctx context.Context, c *config.EverestConfig, l *zap.SugaredLogger) (*EverestServer, error) {
-	options := &cache.Options{
-		ByObject: map[ctrlclient.Object]cache.ByObject{
-			&corev1.Secret{}: {
-				Field: fields.SelectorFromSet(fields.Set{"metadata.name": common.EverestBlocklistSecretName}),
-			},
-		},
-	}
-	kubeConnector, err := kubernetes.NewInCluster(l, ctx, options)
+	kubeConnector, err := kubernetes.NewInCluster(l, ctx, nil)
 	if err != nil {
 		return nil, errors.Join(err, errors.New("failed creating Kubernetes client"))
 	}
@@ -111,7 +102,25 @@ func NewEverestServer(ctx context.Context, c *config.EverestConfig, l *zap.Sugar
 	echoServer.Use(echomiddleware.RateLimiter(echomiddleware.NewRateLimiterMemoryStore(rate.Limit(c.APIRequestsRateLimit))))
 	middleware, store := sessionRateLimiter(c.CreateSessionRateLimit)
 	echoServer.Use(middleware)
+
+	options := &cache.Options{
+		ByObject: map[client.Object]cache.ByObject{
+			&corev1.Secret{}: {
+				Field: fields.SelectorFromSet(fields.Set{"metadata.name": common.EverestBlocklistSecretName}),
+			},
+		},
+	}
+	blocklistClient, err := kubernetes.NewInCluster(l, ctx, options)
+	if err != nil {
+		return nil, err
+	}
+
+	blockList, err := session.NewBlocklist(ctx, blocklistClient, l)
+	if err != nil {
+		return nil, errors.Join(err, errors.New("failed to configure tokens blocklist"))
+	}
 	sessMgr, err := session.New(
+		blockList,
 		session.WithAccountManager(kubeConnector.Accounts()),
 	)
 	if err != nil {
@@ -123,11 +132,6 @@ func NewEverestServer(ctx context.Context, c *config.EverestConfig, l *zap.Sugar
 		return nil, errors.Join(err, errors.New("failed to get OIDC provider config"))
 	}
 
-	blockList, err := session.NewBlocklist(ctx, kubeConnector, l)
-	if err != nil {
-		return nil, errors.Join(err, errors.New("failed to configure tokens blocklist"))
-	}
-
 	e := &EverestServer{
 		config:        c,
 		l:             l,
@@ -135,7 +139,6 @@ func NewEverestServer(ctx context.Context, c *config.EverestConfig, l *zap.Sugar
 		kubeConnector: kubeConnector,
 		sessionMgr:    sessMgr,
 		attemptsStore: store,
-		blocklist:     blockList,
 		oidcProvider:  oidcProvider,
 	}
 	e.echo.HTTPErrorHandler = e.errorHandlerChain()
@@ -452,7 +455,12 @@ func (e *EverestServer) blocklistMiddleWare() (echo.MiddlewareFunc, error) {
 			if skipper(c) {
 				return next(c)
 			}
-			if isBlocked, err := e.blocklist.IsBlocked(c.Request().Context()); err != nil {
+			ctx := c.Request().Context()
+			token, tErr := common.ExtractToken(ctx)
+			if tErr != nil {
+				return tErr
+			}
+			if isBlocked, err := e.sessionMgr.IsBlocked(ctx, token); err != nil {
 				e.l.Error(err)
 				return err
 			} else if isBlocked {
